[autonaming] Move autonaming config into pkg/cmd
Parsing values from configuration into autonaming is better factored out
into the CLI rather than the engine proper.
