+++ pkg/cmd/pulumi/autonaming/json.go
