+++ pkg/resource/autonaming/strategy_test.go
@@ -27,26 +27,26 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 
 	tests := []struct {
 		name                    string
-		options                 globalAutonaming
+		options                 Global
 		wantOptions             *plugin.AutonamingOptions
 		wantDeleteBeforeReplace bool
 		wantErrMsg              string
 	}{
 		{
 			name:        "no config returns no options",
-			options:     globalAutonaming{},
+			options:     Global{},
 			wantOptions: nil,
 		},
 		{
 			name: "default config returns no options",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
+			options: Global{
+				Default: defaultAutonamingConfig,
 			},
 			wantOptions: nil,
 		},
 		{
 			name: "verbatim config enforces logical name",
-			options: globalAutonaming{
+			options: Global{
 				Default: &verbatimAutonaming{},
 			},
 			wantOptions: &plugin.AutonamingOptions{
@@ -58,9 +58,9 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "verbatim config on provider enforces logical name",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"aws": {
 						Default: &verbatimAutonaming{},
 					},
@@ -75,11 +75,11 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "verbatim config on resource enforces logical name",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"aws": {
-						Default: &defaultAutonamingConfig,
+						Default: defaultAutonamingConfig,
 						Resources: map[string]Autonamer{
 							"aws:s3/bucket:Bucket": &verbatimAutonaming{},
 						},
@@ -95,7 +95,7 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "disabled config",
-			options: globalAutonaming{
+			options: Global{
 				Default: &disabledAutonaming{},
 			},
 			wantOptions: &plugin.AutonamingOptions{
@@ -106,9 +106,9 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "disabled config on provider",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"aws": {
 						Default: &disabledAutonaming{},
 					},
@@ -122,9 +122,9 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "disabled config on resource",
-			options: globalAutonaming{
+			options: Global{
 				Default: &verbatimAutonaming{},
-				Providers: map[string]providerAutonaming{
+				Providers: map[string]Provider{
 					"aws": {
 						Default: &verbatimAutonaming{},
 						Resources: map[string]Autonamer{
@@ -141,11 +141,11 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "provider-specific config overrides default",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"aws": {
-						Default: &patternAutonaming{
+						Default: &Pattern{
 							Pattern: "aws-${name}",
 							Enforce: false,
 						},
@@ -161,16 +161,16 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "resource-specific config overrides provider default",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"aws": {
-						Default: &patternAutonaming{
+						Default: &Pattern{
 							Pattern: "aws-${name}",
 							Enforce: false,
 						},
 						Resources: map[string]Autonamer{
-							"aws:s3/bucket:Bucket": &patternAutonaming{
+							"aws:s3/bucket:Bucket": &Pattern{
 								Pattern: "bucket-${name}",
 								Enforce: true,
 							},
@@ -187,21 +187,21 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "invalid resource type returns error",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
+			options: Global{
+				Default: defaultAutonamingConfig,
 			},
 			wantErrMsg: "invalid resource type format: invalid:type",
 		},
 		{
 			name: "unrelated provider and resource configs are ignored",
-			options: globalAutonaming{
-				Default: &defaultAutonamingConfig,
-				Providers: map[string]providerAutonaming{
+			options: Global{
+				Default: defaultAutonamingConfig,
+				Providers: map[string]Provider{
 					"azure": {
 						Default: &disabledAutonaming{},
 					},
 					"aws": {
-						Default: &defaultAutonamingConfig,
+						Default: defaultAutonamingConfig,
 						Resources: map[string]Autonamer{
 							"aws:s3/object:Object": &disabledAutonaming{},
 						},
@@ -213,9 +213,9 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 		},
 		{
 			name: "global config is used if provider does not define a config other than specific resource",
-			options: globalAutonaming{
+			options: Global{
 				Default: &verbatimAutonaming{},
-				Providers: map[string]providerAutonaming{
+				Providers: map[string]Provider{
 					"aws": {
 						Resources: map[string]Autonamer{
 							"aws:s3/object:Object": &disabledAutonaming{},
@@ -241,6 +241,60 @@ func TestGlobalAutonaming_AutonamingForResource(t *testing.T) {
 			got, deleteBeforeReplace := tt.options.AutonamingForResource(urn, nil)
 			assert.Equal(t, tt.wantDeleteBeforeReplace, deleteBeforeReplace)
 			assert.Equal(t, tt.wantOptions, got)
+		})
+	}
+}
+
+func TestGenerateName(t *testing.T) {
+	t.Parallel()
+	urn := urn.New("mystack", "myproject", "", "aws:s3/bucket:Bucket", "myresource")
+	randomSeed := []byte("test seed")
+
+	tests := []struct {
+		name          string
+		pattern       string
+		want          string
+		wantHasRandom bool
+	}{
+		{
+			name:          "hex generation",
+			pattern:       "${name}-${hex(4)}",
+			want:          "myresource-ccf3",
+			wantHasRandom: true,
+		},
+		{
+			name:          "alphanum generation",
+			pattern:       "${name}-${alphanum(5)}",
+			want:          "myresource-uqk8s",
+			wantHasRandom: true,
+		},
+		{
+			name:          "string generation",
+			pattern:       "${name}-${string(6)}",
+			want:          "myresource-qekgoj",
+			wantHasRandom: true,
+		},
+		{
+			name:          "num generation",
+			pattern:       "${num(7)}_${name}",
+			want:          "4080051_myresource",
+			wantHasRandom: true,
+		},
+		{
+			name:          "uuid generation",
+			pattern:       "${uuid}",
+			want:          "ccf35be6-7106-5ccd-784a-fa394fcdb57c",
+			wantHasRandom: true,
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			got, hasRandom := generateName(tt.pattern, urn, randomSeed)
+			assert.Equal(t, tt.want, got)
+			assert.Equal(t, tt.wantHasRandom, hasRandom)
 		})
 	}
 }
